+++ Cargo.toml
@@ -21,7 +21,11 @@ repository = "https://github.com/rust-lang/mdBook"
 rust-version = "1.85.0" # Keep in sync with installation.md and .github/workflows/main.yml
 
 [workspace.dependencies]
+ammonia = "4.1.1"
 anyhow = "1.0.98"
+elasticlunr-rs = "3.0.2"
+handlebars = "6.3.2"
+hex = "0.4.3"
 log = "0.4.27"
 mdbook-core = { path = "crates/mdbook-core" }
 mdbook-html = { path = "crates/mdbook-html" }
@@ -30,10 +34,12 @@ mdbook-preprocessor = { path = "crates/mdbook-preprocessor" }
 mdbook-renderer = { path = "crates/mdbook-renderer" }
 mdbook-summary = { path = "crates/mdbook-summary" }
 memchr = "2.7.5"
+pretty_assertions = "1.4.1"
 pulldown-cmark = { version = "0.10.3", default-features = false, features = ["html"] } # Do not update, part of the public api.
 regex = "1.11.1"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
+sha2 = "0.10.9"
 tempfile = "3.20.0"
 toml = "0.5.11" # Do not update, see https://github.com/rust-lang/mdBook/issues/2037
 
@@ -61,8 +67,6 @@ chrono = { version = "0.4.24", default-features = false, features = ["clock"] }
 clap = { version = "4.3.12", features = ["cargo", "wrap_help"] }
 clap_complete = "4.3.2"
 env_logger = "0.11.1"
-handlebars = "6.0"
-hex = "0.4.3"
 log.workspace = true
 mdbook-core.workspace = true
 mdbook-html.workspace = true
@@ -76,7 +80,6 @@ pulldown-cmark.workspace = true
 regex.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-sha2 = "0.10.8"
 shlex = "1.3.0"
 tempfile.workspace = true
 toml.workspace = true
@@ -95,22 +98,17 @@ tokio = { version = "1.43.1", features = ["macros", "rt-multi-thread"], optional
 axum = { version = "0.8.0", features = ["ws"], optional = true }
 tower-http = { version = "0.6.0", features = ["fs", "trace"], optional = true }
 
-# Search feature
-elasticlunr-rs = { version = "3.0.2", optional = true }
-ammonia = { version = "4.0.0", optional = true }
-
 [dev-dependencies]
 select = "0.6.0"
 semver = "1.0.17"
 snapbox = { version = "0.6.21", features = ["diff", "dir", "term-svg", "regex", "json"] }
-pretty_assertions = "1.3.0"
 walkdir = "2.3.3"
 
 [features]
 default = ["watch", "serve", "search"]
 watch = ["dep:notify", "dep:notify-debouncer-mini", "dep:ignore", "dep:pathdiff", "dep:walkdir"]
 serve = ["dep:futures-util", "dep:tokio", "dep:axum", "dep:tower-http"]
-search = ["dep:elasticlunr-rs", "dep:ammonia", "mdbook-html/search"]
+search = ["mdbook-html/search"]
 
 [[bin]]
 doc = false
