+++ crates/mdbook-html/src/html_handlebars/mod.rs
@@ -1,9 +1,7 @@
-pub use self::hbs_renderer::HtmlHandlebars;
-pub use self::static_files::StaticFiles;
-
 mod hbs_renderer;
 mod helpers;
-mod static_files;
-
 #[cfg(feature = "search")]
 mod search;
+mod static_files;
+
+pub use self::hbs_renderer::HtmlHandlebars;
