+++ crates/mdbook-html/src/lib.rs
@@ -1,3 +1,6 @@
 //! mdBook HTML renderer.
 
+mod html_handlebars;
 pub mod theme;
+
+pub use html_handlebars::HtmlHandlebars;
