+++ src/book/mod.rs
@@ -10,11 +10,14 @@ mod init;
 
 pub use self::book::load_book;
 pub use self::init::BookBuilder;
+use crate::preprocess::{CmdPreprocessor, IndexPreprocessor, LinkPreprocessor};
+use crate::renderer::{CmdRenderer, MarkdownRenderer};
 use anyhow::{Context, Error, Result, bail};
 use log::{debug, error, info, log_enabled, trace, warn};
 pub use mdbook_core::book::{Book, BookItem, BookItems, Chapter, SectionNumber};
 use mdbook_core::config::{Config, RustEdition};
 use mdbook_core::utils;
+use mdbook_html::HtmlHandlebars;
 use mdbook_preprocessor::{Preprocessor, PreprocessorContext};
 use mdbook_renderer::{RenderContext, Renderer};
 pub use mdbook_summary::{Link, Summary, SummaryItem, parse_summary};
@@ -25,9 +28,6 @@ use std::process::Command;
 use tempfile::Builder as TempFileBuilder;
 use toml::Value;
 use topological_sort::TopologicalSort;
-
-use crate::preprocess::{CmdPreprocessor, IndexPreprocessor, LinkPreprocessor};
-use crate::renderer::{CmdRenderer, HtmlHandlebars, MarkdownRenderer};
 
 /// The object used to manage and build a book.
 pub struct MDBook {
